// ui/src/client/models/ErrorBudget.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Athene
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
import {
    ErrorBudgetPair,
    ErrorBudgetPairFromJSON,
    ErrorBudgetPairFromJSONTyped,
    ErrorBudgetPairToJSON,
} from './';

/**
 * 
 * @export
 * @interface ErrorBudget
 */
export interface ErrorBudget {
    /**
     * 
     * @type {Array<ErrorBudgetPair>}
     * @memberof ErrorBudget
     */
    pair: Array<ErrorBudgetPair>;
}

export function ErrorBudgetFromJSON(json: any): ErrorBudget {
    return ErrorBudgetFromJSONTyped(json, false);
}

export function ErrorBudgetFromJSONTyped(json: any, ignoreDiscriminator: boolean): ErrorBudget {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'pair': ((json['pair'] as Array<any>).map(ErrorBudgetPairFromJSON)),
    };
}

export function ErrorBudgetToJSON(value?: ErrorBudget | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'pair': ((value.pair as Array<any>).map(ErrorBudgetPairToJSON)),
    };
}